+++ packages/test-sync/src/commonTest/kotlin/io/realm/kotlin/test/mongodb/common/AppConfigurationTests.kt
@@ -19,6 +19,8 @@ package io.realm.kotlin.test.mongodb.common
 import io.realm.kotlin.internal.platform.PATH_SEPARATOR
 import io.realm.kotlin.internal.platform.appFilesDirectory
 import io.realm.kotlin.internal.platform.runBlocking
+import io.realm.kotlin.log.LogLevel
+import io.realm.kotlin.log.RealmLog
 import io.realm.kotlin.mongodb.App
 import io.realm.kotlin.mongodb.AppConfiguration
 import io.realm.kotlin.mongodb.internal.AppConfigurationImpl
@@ -413,6 +415,21 @@ class AppConfigurationTests {
 //        assertTrue(headerSet.get())
 //        looperThread.testComplete()
 //    }
+
+    @Test
+    fun logLevelDoesNotGetOverwrittenByConfig() {
+        val originalLogLevel = RealmLog.level
+        try {
+            val expectedLogLevel = LogLevel.ALL
+            RealmLog.level = expectedLogLevel
+
+            AppConfiguration.create("")
+
+            assertEquals(expectedLogLevel, RealmLog.level)
+        } finally {
+            RealmLog.level = originalLogLevel
+        }
+    }
 
     @Test
     fun injectedBundleId() {